Displays the set name instead of the setID. Purely aesthetic update.
