+++ src/components/pokemon-grid-display/PokemonGridDisplay.tsx
@@ -22,6 +22,7 @@ import PokemonPolaroidLoading from '@/components/pokemon-cards/pokemon-polaroid/
 import PokemonSet from '@/components/pokemon-cards/pokemon-set/PokemonSet'
 import PokemonSetLoading from '@/components/pokemon-cards/pokemon-set/PokemonSetLoading'
 import { CardSearch } from '@/components/card-filter/CardSearch'
+import { getSetGroups } from '@/utils/pokemonSet'
 
 // Hooks
 import { FiltersProvider } from '@/hooks/useFilters'
@@ -120,53 +121,10 @@ const PokemonGridDisplay = ({ originalPage }: PokemonGridDisplayProps) => {
     })
 
     useEffect(() => {
-        fetch('/api/pokemon-set')
-            .then((res) => res.json())
-            .then((data) => {
-                const sets = Array.isArray(data) ? data : [data]
-
-                const groups: Record<string, PokemonSetType[]> = {
-                    sv: [],
-                    swsh: [],
-                    sm: [],
-                    xy: [],
-                    bw: [],
-                    dp: [],
-                    me: [],
-                    base: [],
-                    other: [],
-                    ex: [],
-                    neo: [],
-                    pl: [],
-                    hgss: [],
-                    pop: [],
-                    ecard: []
-                }
-
-                sets.forEach((set) => {
-                    const id = set.id?.toLowerCase() ?? ''
-                    if (id.includes('sv')) groups.sv.push(set)
-                    else if (id.includes('swsh')) groups.swsh.push(set)
-                    else if (id.includes('sm')) groups.sm.push(set)
-                    else if (id.includes('xy')) groups.xy.push(set)
-                    else if (id.includes('bw')) groups.bw.push(set)
-                    else if (id.includes('hgss') || id.includes('tk-hs'))
-                        groups.hgss.push(set)
-                    else if (id.includes('pl')) groups.pl.push(set)
-                    else if (id.includes('dp')) groups.dp.push(set)
-                    else if (id.includes('me')) groups.me.push(set)
-                    else if (id.includes('ex')) groups.ex.push(set)
-                    else if (id.includes('ecard')) groups.ecard.push(set)
-                    else if (id.includes('pop')) groups.pop.push(set)
-                    else if (id.includes('neo') || id.includes('si'))
-                        groups.neo.push(set)
-                    else if (id.includes('base')) groups.base.push(set)
-                    else groups.other.push(set)
-                })
-
-                setGroupedSets(groups)
-            })
-            .catch((err) => console.error('Error loading sets.json:', err))
+        const fetchGroups = async () => {
+            setGroupedSets(await getSetGroups())
+        }
+        fetchGroups()
     }, [])
 
     const selectStyles = {
